+++ components/UI/FileManagement.js
@@ -1,6 +1,7 @@
 import * as DocumentPicker from "expo-document-picker";
 import { Platform } from "react-native";
 import * as ExpoFileSystem from "expo-file-system";
+import { useState } from "react";
 // import * as Sharing from "expo-sharing";
 
 export async function FileUpload() {
@@ -15,16 +16,39 @@ export async function FileUpload() {
       return;
     }
     console.log(res.assets);
-
     let fileContent = null;
+
+    const setFileContent = (content) => {
+      fileContent = content;
+    };
     if (Platform.OS === "android") {
-      fileContent = await ExpoFileSystem.readAsStringAsync(
-        res.assets[0].uri,
-        { encoding: ExpoFileSystem.EncodingType.UTF8 }
+      setFileContent(
+        await ExpoFileSystem.readAsStringAsync(res.assets[0].uri, {
+          encoding: ExpoFileSystem.EncodingType.UTF8,
+        })
       );
     } else if (Platform.OS === "web") {
       // TODO marc: now read the file using the uri
       console.log("TODO: implement read file for web");
+      const file = res.assets[0].file;
+      console.log(file);
+      if (typeof file === "string") {
+        throw new Error("On web we need to provide a File object");
+      }
+
+      await readFile(file, setFileContent);
+      // const requestData = new FormData();
+      // requestData.append("file", file);
+
+      // const options = {
+      //   method: "POST",
+      //   body: requestData,
+      //   // headers: headers,
+      // };
+
+      // const response = await fetch("", options);
+      // console.log(response);
+      console.log("after reading");
       console.log(fileContent);
     }
     if (fileContent === null || fileContent === undefined) {
@@ -34,7 +58,7 @@ export async function FileUpload() {
     // console.log(fileContent);
 
     try {
-      fileContent = JSON.parse(fileContent);
+      setFileContent(JSON.parse(fileContent));
     } catch (err) {
       console.log("error parsing json");
       console.log(err);
@@ -44,6 +68,17 @@ export async function FileUpload() {
   } catch (err) {
     console.log("error -----", err);
   }
+}
+
+async function readFile(file, setFileContent) {
+  const reader = new FileReader();
+
+  reader.onload = (e) => {
+    const content = e.target.result;
+    setFileContent(content);
+  };
+
+  reader.readAsText(file);
 }
 
 export async function JsonDownload(dataToDownload) {